% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percent.R
\name{add_percent}
\alias{add_percent}
\title{Add Proportion Column}
\usage{
add_percent(df, digits = NULL, multiply = NULL)
}
\arguments{
\item{df}{Dataframe to add proportion column.}

\item{digits}{Number of digits to round proportion. Default set to 1.}

\item{multiply}{Option to multiply proportion by 100. Default set to TRUE.}
}
\value{
Proportion. Use multiply argument to specify if output should be multiplied by 100.
}
\description{
Add proportion column quickly to frequency table. Plays nice with pipe operator following count function.
}
\examples{
library(dplyr)
df <- data.frame(location = c("a","b","c","d","e"), patients = c(10, 12, 6, 20, 2))
df \%>\% add_percent(digits = 1, multiply = TRUE)
}
