% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_groups.R
\name{age_groups}
\alias{age_groups}
\title{Create age groups.}
\usage{
age_groups(age_var, type = "decade")
}
\arguments{
\item{age_var}{Age variable in numeric format.}

\item{type}{Specify which preset to use.}
}
\value{
Age groups as factor. Useful for maintaining proper order when plotting or sorting.
}
\description{
Add age groups column using common age grouping presets. R Markdown documentation has list of presets.
}
\examples{
df <- data.frame(Age = sample(0:99, 100))
df$agegrps <- age_groups(df$Age, type = "covid")
}
