% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_suppression.R
\name{apply_suppression}
\alias{apply_suppression}
\title{Data Suppression for Low Category Frequencies}
\usage{
apply_suppression(n, threshold = NULL, replace_with = NULL)
}
\arguments{
\item{n}{Input field to be suppressed.}

\item{threshold}{Threshold when checking for suppression.}

\item{replace_with}{Value to use when input field is suppressed. String, numeric, or NA.}
}
\value{
Suppressed value to be used in public reporting.
}
\description{
Data Suppression for Low Category Frequencies
}
\examples{
df <- data.frame(Counts = c(5, 1, 10, 3, 12, 9, 4))
df$Counts_Suppressed <- apply_suppression(df$Counts, threshold = 5, replace_with = "**")
}
