% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cair2.R
\name{baby_name}
\alias{baby_name}
\title{Remove Baby-variant name from CAIR2}
\usage{
baby_name(name)
}
\arguments{
\item{name}{First name field from CAIR2.}
}
\value{
Clean first name with baby and variants removed.
}
\description{
Remove Baby-variant name from CAIR2
}
\examples{
baby_name("BABYBOY")

baby_name("Twin Girl")
}
