% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_load.R
\name{batch_load}
\alias{batch_load}
\title{Batch Load CSV Files}
\usage{
batch_load(file_names, col_names = FALSE)
}
\arguments{
\item{file_names}{Vector of files with full path names.}

\item{col_names}{Boolean TRUE/FALSE. Use FALSE if your files do not have column headers/names.}
}
\value{
Dataframe of binded files.
}
\description{
Intended to batch load VRBIS death files for analysis. All columns are read in as character to minimize errors.
}
\examples{
\dontrun{
files <- list.files(path = "G:/file_path/Files/", full.names = TRUE, pattern = ".csv")[39:40]
df <- batch_load(files, col_names = FALSE)
}
}
