% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_load_vrbis.R
\name{batch_load_vrbis}
\alias{batch_load_vrbis}
\title{Batch Load Vital Records}
\usage{
batch_load_vrbis(file_names)
}
\arguments{
\item{file_names}{Vector of files with full path names.}
}
\value{
Dataframe of binded files.
}
\description{
Intended to batch load VRBIS death files for analysis. All columns are read in as character to minimize errors when binding. Assumes all files have same number of columns.
}
\examples{
files_to_load <-list.files(path = "G:/your_file_path/", pattern = "^death", full.names = TRUE)
merged_data <- batch_load_vrbis(files_to_load)
}
