% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cd_cols}
\alias{cd_cols}
\title{Complete list of colors.}
\format{
An object of class \code{character} of length 8.
}
\usage{
cd_cols
}
\description{
Complete list of colors.
}
\keyword{datasets}
