% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cdcd_color}
\alias{cdcd_color}
\title{Use Team Colors in Plots}
\usage{
cdcd_color(...)
}
\arguments{
\item{...}{Name of color(s) from cd_cols.}
}
\value{
Hex codes.
}
\description{
Function to call commonly used colors for data visualizations.
}
\details{
Use names(cd_cols) to return all possible colors.
}
\examples{
cdcd_color("green")
}
