% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_city}
\alias{clean_city}
\title{Find Closest City Match}
\usage{
clean_city(x, reference, threshold = 0.15, redcap = FALSE, ooc = TRUE)
}
\arguments{
\item{x}{Character, input city data.}

\item{reference}{Character, provide your own vector of correctly spelled city names for reference. For OC users, leave blank to use internal vector of OC city names.}

\item{threshold}{Numeric, range 0 to 1. Controls sensitivity/specificity of string distance matching. Default set to 0.15.}

\item{redcap}{Logical, convert character to numeric mapping for REDCap bulk upload.}

\item{ooc}{Logical, choice to keep or recode out of county cities to NA.}
}
\value{
Character, returned city names with closest match to reference vector.
}
\description{
Instead of hard coding spelling on city names, use string distance to find the closest match and recode. For OC users, a reference list of correctly spelled city names have been included, as well as hard coding of common abbreviations/nicknames. For non-OC users, you will need to provide a vector or data.frame of correctly spelled city names for string distance matching.
}
\examples{
clean_city("Anahim")
clean_city("Anahim", redcap = TRUE)
}
