% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_city_match.R
\name{closest_city_match}
\alias{closest_city_match}
\title{Find Closest City Match}
\usage{
closest_city_match(city, possible_matches, threshold = 0.15)
}
\arguments{
\item{city}{Patient/record city variable.}

\item{possible_matches}{List of clean city names for matching. Can be a vector or dataframe column.}

\item{threshold}{Numeric value ranging 0 to 1. Default set to 0.15.}
}
\value{
Column with closest match.
}
\description{
Find the closest matching city name by calculating string distance and selecting results when distance is below threshold. String distance calculated using Jaro-Winkler distance method, which measures edit distance or the number of operations required to transform one string to another. Function requires both stringdist package and a vector or dataframe of clean Orange County cities.
}
\details{
Common city abbreviations/neighborhood names are fixed prior to string comparison i.e. Capo Beach/Capistrano Beach recoded to Dana Point.
}
\examples{
library(dplyr)
df <- data.frame(City = c("Anahem","El Toro","Hntington Bch","Westminister"))

df <- df \%>\% rowwise() \%>\% mutate(recoded_city = closest_city_match(City, oc_cities))
}
