% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{combine_redcap_checkboxes}
\alias{combine_redcap_checkboxes}
\title{Unite Checkbox Variables in REDCap}
\usage{
combine_redcap_checkboxes(df, prefix = NULL, sep = ", ", drop_cols = FALSE)
}
\arguments{
\item{df}{Data.frame or tibble.}

\item{prefix}{Character, prefix of checkbox variable to be combined. If left blank or unspecified, all checkbox variables will be impacted.}

\item{sep}{Character, delimiter to separate multiple values.}

\item{drop_cols}{Logical, option to drop united columns after transformation. Default set to FALSE.}
}
\value{
Data.frame or tibble with united colums using prefix.
}
\description{
While useful for data collection, checkbox variables may need to be merged/united into a singular variable for analysis. REDCap internally labels checkbox variables are "prefix___1", with the number matchin the number response in the checkbox variable. Warning: if REDCap ever changes this format, this function will stop working.
}
