% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dates.R
\name{complete_dates}
\alias{complete_dates}
\title{Complete Time Series}
\usage{
complete_dates(df, start_date = NULL, level = c("day", "week", "month"))
}
\arguments{
\item{df}{Dataframe of initial time series data.}

\item{start_date}{Specify date to start time series. Last date in series is oldest/most recent date.}

\item{level}{Options to complete time series at the day, week, or month level. When using week, be mindful you are completing sequence using MMWR week ending date. For month, works best when date is M/1/YYYY.}
}
\value{
Dataframe of complete time series starting at specified start date.
}
\description{
Complete time series data at the day, week, or month level starting at a specified date. Useful for filling in missing dates when preparing for plotting or export to Tableau.
}
\examples{
df <- complete_dates(df, start_date = min(df$Dates), level = "month")
}
