% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{create_unique_ids}
\alias{create_unique_ids}
\title{Create Unique Alphanumeric ID's}
\usage{
create_unique_ids(n, char_len = 12)
}
\arguments{
\item{n}{Numeric or integer, how many ID's do you need.}

\item{char_len}{Numeric or integer, total ID character length. Default set to 12.}
}
\value{
Character, length of n unique ID's.
}
\description{
Create Unique Alphanumeric ID's
}
\examples{
create_unique_ids(n = 5)
}
