% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{dashboard_tbl}
\alias{dashboard_tbl}
\title{Aggregate Data for Dashboard Reporting}
\usage{
dashboard_tbl(
  data,
  group_by = NULL,
  reverse = TRUE,
  digits = 1,
  n_suppress = NULL
)
}
\arguments{
\item{data}{Input dataframe containing only those columns that need to be summarized.}

\item{group_by}{To be used if summarizing data over multiple years or groups.}

\item{reverse}{Sets order of n and percent in label.}

\item{digits}{Number of digits to round proportion.}

\item{n_suppress}{Suppress values less than specified value.}
}
\value{
Dataframe with n, proportion, and label.
}
\description{
Summarize data across columns in preparation for dashboard reporting.
}
