% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{delete_redcap_record}
\alias{delete_redcap_record}
\title{Delete a Record in REDCap}
\usage{
delete_redcap_record(
  url,
  token,
  record_id,
  instrument = NULL,
  repeat_instance = NULL
)
}
\arguments{
\item{url}{Character, API url.}

\item{token}{Character, token from project.}

\item{record_id}{Numeric, single or vector of record id's.}

\item{instrument}{Character, name of instrument that contains data you wish to delete. Optional.}

\item{repeat_instance}{Character, repeating instance number that contains data you wish to delete. Optional.}
}
\value{
Number of records deleted  or (if instrument or repeating instance is provided) the number of items deleted over the total records specified.
}
\description{
This method allows you to delete one or more records from a project, and also optionally allows you to delete parts of records such as specific instruments and/or repeating instances.
}
