% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desaturate_geom.R
\name{desaturate_geom}
\alias{desaturate_geom}
\alias{ggplot_add.desaturate}
\title{Desaturate Geom Using Expression}
\usage{
desaturate_geom(
  expr,
  pal = NULL,
  size = 3.14,
  desaturate = 0.75,
  linewidth = 1.2
)
}
\arguments{
\item{expr}{Expression}

\item{pal}{Color for highlighted data point}

\item{size}{Point size if using geom_point}

\item{desaturate}{Level of desaturation for non-highlighted data points, ranges from 0-1 with 1 equal to highest level of desaturation}

\item{linewidth}{Linewidth if using geom_line or geom_sf}
}
\value{
Desaturated geon
}
\description{
Highlight a bar, point, or line plot based on an expression while desaturating all other data points.
}
\examples{
df <- data.frame(locations = letters[1:5], scores = c(80,84,91,89,80))
ggplot(data = df, aes(x = locations, y = scores)) +
geom_col() +
desaturate_geom(scores==max(scores), pal = "#d55c19", desaturate = 0.7)
}
