% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elr.R
\name{pos}
\alias{pos}
\alias{neg}
\title{ELR Terms - Positive}
\usage{
pos(collapse = FALSE)

neg(collapse = FALSE)
}
\arguments{
\item{collapse}{Boolean to collapse string separated by "|".}
}
\value{
Vector of characters.

Vector of characters.
}
\description{
A vector of SNOMED codes and keywords to be used with ELR recoding.

A vector of SNOMED codes and keywords to be used with ELR recoding.
}
\examples{
pos(collapse = FALSE)
pos()
neg(collapse = FALSE)
neg()
}
