% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_points.R
\name{end_points}
\alias{end_points}
\title{Subset Data To Label Line Plot}
\usage{
end_points(df, date_col)
}
\arguments{
\item{df}{Dataframe}

\item{date_col}{Date column to be used in plot - can be date, year, or week.}
}
\value{
Dataframe filtered at maximum (or most recent) date including all groups at that date.
}
\description{
Subset dataframe to label last point in geom_line.
}
\examples{
df <- data.frame(Date = seq.Date(from = as.Date("2023-01-01"), to = as.Date("2023-12-01"),
by = "month"), scores = sample(65:99, 12))
end_points(df, Date)
}
