% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cleveland.R
\name{geom_cleveland}
\alias{geom_cleveland}
\title{Cleveland Dot Plot}
\usage{
geom_cleveland(
  data = NULL,
  mapping = NULL,
  show.legend = NA,
  inherit.aes = TRUE,
  position = "identity",
  na.rm = FALSE,
  colour_x = NULL,
  colour_xend = NULL,
  size_x = NULL,
  linewidth = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe.}

\item{mapping}{Requires following aesthetics: x = starting point, xend = ending point, y = category/factor.}

\item{show.legend}{Boolean.}

\item{inherit.aes}{If TRUE combine with default mapping at the top level of the plot.}

\item{position}{Default set to "identity".}

\item{na.rm}{Boolean to remove NA values.}

\item{colour_x}{Color of starting point.}

\item{colour_xend}{Color of ending point.}

\item{size_x}{Size of points.}

\item{linewidth}{Line thickness.}

\item{...}{Other args, man.}
}
\value{
Cleveland dot plot.
}
\description{
Cleveland dot plot (or dumbbell plot) using ggplot2. Arose out of need to create timelines.
}
\examples{
library(ggplot2)
df <- data.frame(Country = c("UK-A","US-A","UK-B","US-B"), PRE = c(0,1,2,3), POST = c(4,5,6,7))
ggplot(data = df, aes(x = PRE, xend = POST, y = Country)) +
 geom_cleveland(colour_x = "blue", colour_xend = "red") +
 labs(title = "Cleveland Dot Plot", x = "Pre/Post Scores", y = "Country")
}
