% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dumbbell.R
\docType{data}
\name{geom_dumbbell}
\alias{geom_dumbbell}
\alias{GeomDumbbell}
\title{Dumbbell Plot}
\format{
An object of class \code{GeomDumbbell} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
geom_dumbbell(
  data = NULL,
  mapping = NULL,
  show.legend = NA,
  inherit.aes = TRUE,
  position = "identity",
  na.rm = FALSE,
  colour_x = NULL,
  colour_xend = NULL,
  size = NULL,
  linewidth = NULL,
  colour = NULL,
  ...
)

GeomDumbbell
}
\arguments{
\item{data}{Dataframe.}

\item{mapping}{Requires following aesthetics: x = starting point, xend = ending point, y = category/factor.}

\item{show.legend}{Boolean.}

\item{inherit.aes}{If TRUE combine with default mapping at the top level of the plot.}

\item{position}{Default set to "identity".}

\item{na.rm}{Boolean to remove NA values.}

\item{colour_x}{Color of starting point.}

\item{colour_xend}{Color of ending point.}

\item{size}{Size of points.}

\item{linewidth}{Line thickness.}

\item{colour}{Line color.}

\item{...}{Other args, man.}
}
\value{
Dumbbell plot.
}
\description{
Line plot with two points at each end representing different values.
}
\examples{
library(ggplot2)
df <- data.frame(Country = c("UK-A","US-A","UK-B","US-B"), PRE = c(0,1,2,3), POST = c(4,5,6,7))
ggplot(data = df, aes(x = PRE, xend = POST, y = Country)) +
 geom_dumbbell(colour_x = "blue", colour_xend = "red", size = 3, linewidth = 3, colour = "black") +
 labs(title = "Dumbbell Plot", x = "Pre/Post Scores", y = "Country")
}
\keyword{datasets}
