% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_geom.R
\name{highlight_geom}
\alias{highlight_geom}
\title{Highlight Geom Using Expression}
\usage{
highlight_geom(expr, pal, size = 5, alpha = 0.5, linewidth = 1.5)
}
\arguments{
\item{expr}{Expression}

\item{pal}{color for highlighted geom}

\item{size}{Point size if using geom_point}

\item{alpha}{Alpha if using geom_point}

\item{linewidth}{Linewidth if using geom_line}
}
\value{
ggplot2 with highlighted layer
}
\description{
Highlight a bar, point, or line plot based on an expression.
}
\examples{
df <- data.frame(locations = letters[1:5], scores = c(80,84,91,89,80))
ggplot(data = df, aes(x = locations, y = scores)) +
geom_col() +
highlight_geom(scores==max(scores), pal = "#9E0059")
}
