% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_title.R
\name{highlight_title}
\alias{highlight_title}
\title{Highlight Keywords in ggplot2 Title}
\usage{
highlight_title(text = NULL, words = NULL)
}
\arguments{
\item{text}{Title, string.}

\item{words}{Named vector with word(s) and hex code.}
}
\value{
Title with highlighted word/words.
}
\description{
When building ggplot using highlight_geom, you may need to color a word in the title to draw focus. Function depends on ggtext::element_markdown() to be added to theme plot.
}
\examples{
\dontrun{
df <- data.frame(locations = letters[1:5], events = c(20,24,19,12,33))
ggplot(data = df, aes(x = locations, y = events)) +
geom_col() +
 labs(
   title = highlight_title("New Title Goes Here FRIEND", words = c("New"="#d55c19"))
 ) +
 theme(
   plot.title = ggtext::element_markdown()
 )
 }
}
