% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_redact.R
\name{hiv_redact}
\alias{hiv_redact}
\title{Redact HIV/AIDS from Dataframe}
\usage{
hiv_redact(df)
}
\arguments{
\item{df}{Dataframe to scan.}
}
\value{
Dataframe with HIV and/or AIDS removed.
}
\description{
Remove iterations of HIV and/or AIDS from dataframe.
}
\examples{
df <- data.frame(cause = c("cancer","hepatitis","COVID-19","HIV"))
hiv_redact(df)
}
