% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_id_1}
\alias{match_id_1}
\title{Match ID #1}
\usage{
match_id_1(first_name, last_name, dob, part_address)
}
\arguments{
\item{first_name}{First 4 characters of string}

\item{last_name}{First 4 characters of string}

\item{dob}{Date}

\item{part_address}{First 10 characters of string}
}
\value{
Match id using part of name, date of birth, and the first 10 characters of the address.
}
\description{
Match ID for linking disparate datasets using parts of identifiers. Be aware of address formatting and consider recoding variable using clean_address.
}
\examples{
match_id_1("Mickey","Mouse","1955-07-17","1313 Disneyland Dr")
}
