% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_id_2}
\alias{match_id_2}
\title{Match ID #2}
\usage{
match_id_2(first_name, last_name, dob, address)
}
\arguments{
\item{first_name}{First 4 characters of string}

\item{last_name}{First 4 characters of string}

\item{dob}{Date}

\item{address}{Full address}
}
\value{
Match id using part of name, date of birth, and full address.
}
\description{
Match ID for linking disparate datasets using parts of identifiers.
}
\examples{
match_id_2("Mickey","Mouse","1955-07-17","1313 Disneyland Dr")
}
