% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_id_3}
\alias{match_id_3}
\title{Match ID #3 - Partial Name, DOB, Phone Number}
\usage{
match_id_3(first_name, last_name, dob, phone_number)
}
\arguments{
\item{first_name}{First 4 characters of string}

\item{last_name}{First 4 characters of string}

\item{dob}{Date}

\item{phone_number}{String or Numeric}
}
\value{
Match id using part of name, date of birth, and phone number.
}
\description{
Match ID for linking disparate datasets using parts of identifiers. Be aware of phone number formatting and consider recoding variable using clean_phone.
}
\examples{
match_id_3("Mickey","Mouse","1955-07-17","7147814636")
}
