% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_id_4}
\alias{match_id_4}
\title{Match ID #4 - Partial Name, DOB}
\usage{
match_id_4(first_name, last_name, dob)
}
\arguments{
\item{first_name}{First 4 characters of string}

\item{last_name}{First 4 characters of string}

\item{dob}{Date}
}
\value{
Match id using part of name and date of birth.
}
\description{
Match ID for linking disparate datasets using parts of identifiers.
}
\examples{
match_id_4("Mickey","Mouse","1955-07-17")
}
