% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr.R
\name{mmwr_calendar}
\alias{mmwr_calendar}
\title{Create MMWR Calendar}
\usage{
mmwr_calendar(year)
}
\arguments{
\item{year}{Epidemologic year of interest.}
}
\value{
Dataframe.
}
\description{
Calculate calendar with disease weeks plus start and end dates.
}
\examples{
mmwr_calendar(2024)
}
