% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr.R
\name{mmwrweek_to_date}
\alias{mmwrweek_to_date}
\title{Convert Disease Week to Week Ending Date}
\usage{
mmwrweek_to_date(year, week)
}
\arguments{
\item{year}{MMWR Year.}

\item{week}{MMWR Week.}
}
\value{
Week ending date (Saturday).
}
\description{
When calendar date is unknown, convert disease week to week ending date.
}
\examples{
mmwrweek_to_date(2023, 52)
}
