% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndromic.R
\name{nssp_date_range}
\alias{nssp_date_range}
\title{Adjust NSSP API Date Range}
\usage{
nssp_date_range(url, start = NULL, end = NULL)
}
\arguments{
\item{url}{ESSENCE API URL.}

\item{start}{Start date for data extraction, date format.}

\item{end}{End date for data extraction, date format.}
}
\value{
String API URL with new date range.
}
\description{
Customize date range in API URL without having to recreate within ESSENCE.
}
