% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_city_sf}
\alias{oc_city_sf}
\title{OC City Basemap}
\format{
An object of class sf.
}
\usage{
data(oc_city_sf)
}
\description{
OC City Basemap
}
\examples{
data(oc_city_sf)
}
\keyword{datasets}
