% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\docType{data}
\name{oc_zip_sf}
\alias{oc_zip_sf}
\title{Orange County Zip Codes}
\format{
An sf object with 86 rows and 7 columns.
}
\usage{
oc_zip_sf
}
\description{
A simple features (sf) dataset containing the geometry of Orange County zip codes.
}
\keyword{datasets}
