% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dumbbell.R, R/geom_lollipop.R,
%   R/highlight_geom.R
\docType{data}
\name{GeomDumbbell}
\alias{GeomDumbbell}
\alias{GeomLollipop}
\alias{ggplot_add.highlight}
\title{geom_dumbell proto}
\description{
geom_dumbell proto

geom_lollipop proto

ggplot_add.highlight
}
\keyword{datasets}
