% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_words.R
\name{pretty_words}
\alias{pretty_words}
\title{Messy String to Title Casing}
\usage{
pretty_words(x)
}
\arguments{
\item{x}{String variable.}
}
\value{
Input string in title casing format.
}
\description{
Convert string to title casing. Useful for fixing patient or street names.
}
\examples{
pretty_words("MeSsY DaTa GiVeS mE a HeAdAcHe.")
}
