% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_rate.R
\name{rate_per_100k}
\alias{rate_per_100k}
\title{Calculate Incidence Rate per 100,000}
\usage{
rate_per_100k(n, pop, digits = 1)
}
\arguments{
\item{n}{Number of cases.}

\item{pop}{Population denominator.}

\item{digits}{Number of digits to round rate.}
}
\value{
Crude rate per 100,000.
}
\description{
Calculate incidence rate using your own population denominator.
}
\examples{
rate_per_100k(5, 320000, digits = 1)
}
