% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{read_redcap}
\alias{read_redcap}
\title{Import Data from REDCap Project}
\usage{
read_redcap(url, token, content, raw = FALSE)
}
\arguments{
\item{url}{API url, character.}

\item{token}{API token, character.}

\item{content}{Currently limited to: record, userRole, user.}

\item{raw}{Logical to return raw or labeled data.}
}
\value{
Data.frame with project data.
}
\description{
Import project data from REDCap into R via the API. Import options include raw/labeled data and data type.
}
