% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrbis.R
\name{recode_ctract}
\alias{recode_ctract}
\title{Recode VRBIS Census Tract}
\usage{
recode_ctract(col)
}
\arguments{
\item{col}{Expects "Residence Census Tract 2010" variable from VRBIS dataset.}
}
\value{
Recoded census tract as character.
}
\description{
Remove state and county fips code from census tract for matching/joining.
}
\examples{
recode_ctract("06059099244")
recode_ctract(as.numeric("06059099244"))
}
