% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gender.R
\name{recode_gender}
\alias{recode_gender}
\title{Recode Sex Column in CalREDIE}
\usage{
recode_gender(gender_var, ordered = NULL)
}
\arguments{
\item{gender_var}{"Sex" column in CalREDIE.}

\item{ordered}{Return output as factor. Default set to TRUE.}
}
\value{
Recoded input column.
}
\description{
Made specifically for CaLREDIE UDF data export. Recode abbreviations to full names. Future iterations may be applicable to non-CalREDIE datasets.
}
\examples{
example = c("F","F","TF",NA)
recode_gender(example, ordered = FALSE)
}
