% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calredie.R
\name{recode_orientation}
\alias{recode_orientation}
\title{Recode Sexual Orientation in CalREDIE}
\usage{
recode_orientation(col)
}
\arguments{
\item{col}{CTCIAdtlDemOrient column from CalREDIE.}
}
\value{
Recoded column as character.
}
\description{
Made specifically for CaLREDIE UDF data export. Recode abbreviations to full names.
}
\examples{
recode_orientation("BIS")
recode_orientation(NA)
}
