% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_race.R
\name{recode_race}
\alias{recode_race}
\title{Recode Race/Ethnicity in CalREDIE or CAIR2}
\usage{
recode_race(ethnicity, race)
}
\arguments{
\item{ethnicity}{Patient ethnicity variable.}

\item{race}{Patient race variable.}
}
\value{
Merged race/ethnicity variable.
}
\description{
Simplify race/ethnicity patient data to one variable. Hierarchy defaults to Hispanic/Latinx regardless of reported race. Function can handle one (e.g. CAIR2) or two inputs (e.g. CalREDIE).
}
\examples{
recode_race("Hispanic","Black or African American")
recode_race("Native Hawaiian or Other Pacific Islander")
}
