% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe.ifelse}
\alias{safe.ifelse}
\title{Retain class with ifelse inside a function}
\usage{
safe.ifelse(cond, yes, no)
}
\arguments{
\item{cond}{Condition to test}

\item{yes}{Returned output if true}

\item{no}{Returned output if false}
}
\value{
Tested condition while retaining original class.
}
\description{
Retain class with ifelse inside a function
}
\keyword{internal}
