% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Split and Print Dataframe}
\usage{
split_df(path = NULL, df, chunks = NULL, prefix = "List_")
}
\arguments{
\item{path}{File path for file export}

\item{df}{Dataframe to export}

\item{chunks}{Number of rows to split dataframe into}

\item{prefix}{Default set to "List". Option to specify file prefix.}
}
\value{
Multiple .CSV files split into number of rows specified from chunks argument.
}
\description{
Divide a dataframe into multiple .CSV files with a specified number of rows per file.
}
\examples{

split_df(path = "G:/file_path/", df = test_data, chunks = 200, prefix = "list_")
}
