% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supress.R
\name{suppress}
\alias{suppress}
\title{Data Suppression for Low Category Frequencies}
\usage{
suppress(n, threshold = 5, replace_with = "**")
}
\arguments{
\item{n}{Input field to be suppressed.}

\item{threshold}{Threshold when checking for suppression.}

\item{replace_with}{Value to use when input field is suppressed. String, numeric, or NA.}
}
\value{
Suppressed value.
}
\description{
Mask values/counts below a specified threshold to maintain patient confidentiality.
}
\examples{
df <- data.frame(Counts = c(5, 1, 10, 3, 12, 9, 4))
df$Counts_Suppressed <- apply_suppression(df$Counts, threshold = 5, replace_with = "**")
}
