% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_between.R
\name{time_between}
\alias{time_between}
\title{Calculate Time Between Two Dates}
\usage{
time_between(date2, date1, unit = c("days", "weeks", "months", "years"))
}
\arguments{
\item{date2}{More recent date.}

\item{date1}{Older date or date further in the past.}

\item{unit}{Specify if returned time is in days, weeks, months, or years.}
}
\value{
Time interval in specified units.
}
\description{
Calculate the elapsed time between two dates in a specified unit of time. Order of dates within does not necessarily matter, but order will dictate sign of returned output.
}
\examples{
date2 = as.POSIXlt("2023-02-08 00:10:00 PST")
date1 = as.POSIXlt("1986-01-08 22:00:00 PST")

time_between(date2, date1, unit = "days")
}
