% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mmwr_date.R
\name{to_mmwr_date}
\alias{to_mmwr_date}
\title{MMWR Date from Record Date}
\usage{
to_mmwr_date(x, type = c("week", "year", "both"))
}
\arguments{
\item{x}{Input date i.e. episode date.}

\item{type}{Specify to return disease week, year, or both.}
}
\value{
MMWR date for time series aggregation.
}
\description{
Following CDC MMWR schema, calculate week number or year based on input date.
}
\details{
From CDC: The first day of any MMWR week is Sunday. MMWR week numbering is sequential beginning with 1 and
incrementing with each week to a maximum of 52 or 53. MMWR week #1 of an MMWR year is the first week of
the year that has at least four days in the calendar year. For example, if January 1 occurs on a Sunday, Monday,
Tuesday or Wednesday, the calendar week that includes January 1 would be MMWR week #1. If January 1
occurs on a Thursday, Friday, or Saturday, the calendar week that includes January 1 would be the last MMWR
week of the previous year (#52 or #53). Because of this rule, December 29, 30, and 31 could potentially fall into
MMWR week #1 of the following MMWR year.
}
\examples{
episode_date = as.Date("2020-01-14")
to_mmwr_date(episode_date, type = "week")
}
