% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_month.R
\name{to_month}
\alias{to_month}
\title{Date to Month Date}
\usage{
to_month(x)
}
\arguments{
\item{x}{Input date.}
}
\value{
Date in YYYY-MM-01 format.
}
\description{
Recode date to standard month date for dashboards or ggplots2.
}
\examples{
patient_encounter = as.Date("2020-01-14")
to_month(patient_encounter)
}
