% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{to_quarter}
\alias{to_quarter}
\title{Date to Quarter}
\usage{
to_quarter(x, fiscal = FALSE)
}
\arguments{
\item{x}{Input date.}

\item{fiscal}{Boolean, TRUE returns fiscal quarter.}
}
\value{
Character in format YYYY-Q0#.
}
\description{
Recode date to calendar or fiscal quarter.
}
\examples{
to_quarter(Sys.Date())
to_quarter(Sys.Date(), fiscal = TRUE)
}
