% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmwr.R
\name{total_weeks}
\alias{total_weeks}
\title{Calculate Total Disease Weeks for Year}
\usage{
total_weeks(year)
}
\arguments{
\item{year}{Epidemologic year.}
}
\value{
Number of MMWR Disease Weeks.
}
\description{
Calculate total diseases for any given MMWR year (52 or 53 weeks).
}
\examples{
year_start(2024)
}
