% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrbis.R
\name{vrbis_manner_death}
\alias{vrbis_manner_death}
\title{Recode VRBIS Manner of Death}
\usage{
vrbis_manner_death(col)
}
\arguments{
\item{col}{Expects "Manner of Death" column from VRBIS dataset.}
}
\value{
Recoded response as character.
}
\description{
Convert single letter to full responses per CCDF data dictionary.
}
\examples{
vrbis_manner_death("A")
vrbis_manner_death(NA)
}
