% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrbis.R
\name{vrbis_place_death}
\alias{vrbis_place_death}
\title{Recode VRBIS Place of Death}
\usage{
vrbis_place_death(col)
}
\arguments{
\item{col}{Expects "Place of Death (Facility)" column from VRBIS dataset.}
}
\value{
Recoded response as character.
}
\description{
Convert number to full responses per CCDF data dictionary.
}
\examples{
vrbis_place_death(6)
vrbis_place_death("6")
}
