% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrbis.R
\name{vrbis_resident}
\alias{vrbis_resident}
\title{Algorithm for Death Certificate Inclusion}
\usage{
vrbis_resident(
  death_location,
  county_of_death,
  county_fips,
  fips = "059",
  county_code = "30"
)
}
\arguments{
\item{death_location}{Expects "Place of Death (Facility)" column from VRBIS dataset.}

\item{county_of_death}{Expects "County of Death (Code)" column.}

\item{county_fips}{Expects "Decedents County of Residence (NCHS Code)" column.}

\item{fips}{County fips code of local health jurisdiction.}

\item{county_code}{State county code per Appendix G of CCDF data dictionary.}
}
\value{
0/1, with 1 indicating death belongs to local health jurisdiction.
}
\description{
Standardized method to determine if death belongs to local health jurisdiction.
}
