% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_ending_date.R
\name{week_ending_date}
\alias{week_ending_date}
\title{Find Week Ending Date}
\usage{
week_ending_date(x)
}
\arguments{
\item{x}{Input date.}
}
\value{
Output date.
}
\description{
Calculate week ending date following CDC MMWR schema.
}
