% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_labels.R
\name{wrap_labels}
\alias{wrap_labels}
\title{Wrap Long Axis Labels}
\usage{
wrap_labels(x, delim)
}
\arguments{
\item{x}{Input list of labels.}

\item{delim}{Vector of characters or words to break label.}
}
\value{
Wrapped labels.
}
\description{
Break long axis labels at one or more "delimiters".
}
\examples{
df <- data.frame(group = c("Native Hawaiian or Other Pacific Islander",
"Black or African American","American Indian or Alaska Native"), score = c(89.5, 84, 73))

ggplot(data = df, aes(x = group, y = score)) +
 geom_col() +
 scale_x_discrete(labels = wrap_labels(delim = "or"))
}
