% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{write_redcap}
\alias{write_redcap}
\title{Write New or Modified Data to REDCap Project}
\usage{
write_redcap(
  df,
  url,
  token,
  forceAutoNumber = FALSE,
  batch = FALSE,
  batch_delay = 0.5
)
}
\arguments{
\item{df}{Data.frame to write back to REDCap.}

\item{url}{Character, API url.}

\item{token}{Character, token from project.}

\item{forceAutoNumber}{Logical, if TRUE new record ids will be automatically determined.}

\item{batch}{Logical, TRUE if uploading large volumes of data.}

\item{batch_delay}{Numeric, time delay between batches when batch uploading.}
}
\value{
Success or warning message, no data returned.
}
\description{
This method allows you to upload/write a set of records for a project. To use this method, you must have API Export privileges in the project.
}
