% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{year_start}
\alias{year_start}
\title{MMWR Year Start}
\usage{
year_start(year)
}
\arguments{
\item{year}{Year}
}
\value{
Day one of week one following MMWR schema
}
\description{
MMWR Year Start
}
