% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percent.R
\name{add_percent}
\alias{add_percent}
\title{Add Proportion}
\usage{
add_percent(n, digits = 2, multiply = TRUE)
}
\arguments{
\item{n}{Count/frequency variable.}

\item{digits}{Number of digits to round.}

\item{multiply}{Boolean TRUE/FALSE to multiply proportion by 100.}
}
\value{
Proportion.
}
\description{
Calculate percentage per category with options to round and multiply by 100.
}
\examples{
library(dplyr)
starwars \%>\% head(20) \%>\% count(species) \%>\% mutate(percent = add_percent(n))
}
