% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percent.R
\name{add_percent}
\alias{add_percent}
\title{Add Proportion}
\usage{
add_percent(n, digits = 2, multiply = TRUE)
}
\arguments{
\item{n}{Count/frequency variable.}

\item{digits}{Number of digits to round.}

\item{multiply}{Default set to TRUE. To keep as fraction set to FALSE.}
}
\value{
Proportion.
}
\description{
Calculate percentage per category with options to round and multiply by 100.
}
\examples{
df <- data.frame(species = c("Droid","Human","Wookie"), n = c(3,12,1))
df$percent <- add_percent(df$n)
df
}
