% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apollo.R
\name{apollo_label}
\alias{apollo_label}
\title{Apollo Label}
\usage{
apollo_label(direction = c("vertical", "horizontal"), ...)
}
\arguments{
\item{direction}{If using coord_flip, specify "horizontal". Otherwise use "vertical".}

\item{...}{Other arguments available from geom_text.}
}
\value{
Standard label for plots.
}
\description{
Apollo Label
}
