% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apollo.R
\name{apollo_label}
\alias{apollo_label}
\title{Apollo Label}
\usage{
apollo_label(..., size = 4.5, color = "#231f20")
}
\arguments{
\item{...}{Other arguments available from geom_text.}

\item{size}{Font size, preset to 4.5}

\item{color}{Font color, preset to #231f20}
}
\value{
Standard label for plots.
}
\description{
Labels to match the style in theme_apollo.
}
\details{
Recommended settings:
bar plot vjust = -0.3
horizontal bar plot hjust = -0.5.
}
