% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_season.R
\name{assign_season}
\alias{assign_season}
\title{Assign Respiratory Season}
\usage{
assign_season(date)
}
\arguments{
\item{date}{Input date.}
}
\value{
Character in format XXXX-XX.
}
\description{
Assign Respiratory Season
}
\examples{
x = as.Date("2023-10-01")
assign_season(x)
}
