% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cair2.R
\name{baby_name}
\alias{baby_name}
\title{Remove Baby Name in CAIR2}
\usage{
baby_name(name)
}
\arguments{
\item{name}{First name field from CAIR2.}
}
\value{
Clean first name.
}
\description{
Clean up birth hepatitis B doses in CAIR2.
}
\examples{
baby_name("BABYBOY")
baby_name("Twin Girl")
}
