% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cdcd_color}
\alias{cdcd_color}
\title{Use Team Colors in Plots}
\usage{
cdcd_color(...)
}
\arguments{
\item{...}{Vector of color names. Run cdcd_color() to view all options.}
}
\value{
Hex codes.
}
\description{
Function to call commonly used colors for data visualizations.
}
\examples{
cdcd_color()
cdcd_color("green")
}
