% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean patient/record address}
\usage{
clean_address(address_var, keep_extra = NULL)
}
\arguments{
\item{address_var}{Address variable.}

\item{keep_extra}{Option to keep/remove extra addreses (e.g. Apartment, Unit, Space). TRUE/FALSE, default set to TRUE.}
}
\value{
Cleaned up address in character format.
}
\description{
Recode common address issues such as cardinal directions and street names, and removes extra location information such as unit or apartment number. Improves results during batch geocoding or matching to other datasets.
}
\examples{
x = "1234 N Main St Apt 405"
clean_address(x, keep_extra = TRUE)
}
