% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean patient/record address}
\usage{
clean_address(address_var, keep_extra = TRUE)
}
\arguments{
\item{address_var}{Address variable.}

\item{keep_extra}{Defaults to TRUE. Option to keep/remove extra address components (e.g. Apartment, Unit, Space).}
}
\value{
Address as character.
}
\description{
Recode common address issues such as cardinal directions and street names, and removes extra location information such as unit or apartment number.
}
\examples{
x = "1234 N Main St Apt 405"
clean_address(x, keep_extra = TRUE)
clean_address(x, keep_extra = FALSE)
}
