% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calredie.R
\name{clean_phone}
\alias{clean_phone}
\title{Clean Phone Number}
\usage{
clean_phone(x)
}
\arguments{
\item{x}{Phone Character or numeric, number variable.}
}
\value{
Character, phone number containing 10 digits if valid.
}
\description{
Reformat phone number column to 10 digit U.S. format. If international or invalid phone number (e.g. not 10 digits), returns NA.
}
\examples{
clean_phone("1-714-834-8180")
}
