% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elr.R
\name{pos}
\alias{pos}
\alias{neg}
\title{ELR Terms}
\usage{
pos(collapse = FALSE, add_in = NULL)

neg(collapse = FALSE, add_in = NULL)
}
\arguments{
\item{collapse}{Boolean to collapse string separated by "|".}

\item{add_in}{Vector of additional lab results.}
}
\value{
Vector of characters.
}
\description{
A vector of SNOMED codes and keywords to be used with ELR recoding.
}
\examples{
pos(collapse = TRUE)
pos()
neg(collapse = TRUE)
neg()
}
