% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_points.R
\name{end_points}
\alias{end_points}
\title{Subset Data To Label Line Plot}
\usage{
end_points(df, date = NULL, group_by = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{date}{Date column to be used in plot - can be date, year, or week.}

\item{group_by}{For time series that end at different dates, specify group variable.}
}
\value{
Data.frame filtered at maximum (or most recent) date including all groups at that date.
}
\description{
Subset dataframe to label last point in geom_line. Function takes into account time series that don't end at the same time point.
}
\examples{
df <- data.frame(Date = seq.Date(from = as.Date("2023-01-01"), to = as.Date("2023-12-01"),
by = "month"), scores = sample(65:99, 12))
end_points(df, Date)
}
