% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lollipop.R
\name{geom_lollipop}
\alias{geom_lollipop}
\title{Lollipop Plot}
\usage{
geom_lollipop(
  data = NULL,
  mapping = NULL,
  show.legend = NA,
  inherit.aes = TRUE,
  position = "identity",
  na.rm = FALSE,
  linewidth = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe}

\item{mapping}{x and y mapping.}

\item{show.legend}{Boolean.}

\item{inherit.aes}{If TRUE combine with default mapping at the top level of the plot.}

\item{position}{Default set to "identity".}

\item{na.rm}{Boolean to remove NA values.}

\item{linewidth}{Change linewidth.}

\item{...}{The other arguments!}
}
\value{
Lollipop plot.
}
\description{
Point plot with stem.
}
\examples{
df <- data.frame(locations = letters[1:5], scores = c(90,82,87,91,74))

ggplot(data = df, aes(x = locations, y = scores)) +
 geom_lollipop(size = 10, linewidth = 2, color = "red")
}
