% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linelist}
\alias{linelist}
\title{Outbreak Linelist}
\format{
A tibble containing no PHI.
}
\usage{
data(linelist)
}
\description{
Outbreak Linelist
}
\examples{
data(linelist)
}
\keyword{datasets}
