% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_percent.R
\name{n_percent}
\alias{n_percent}
\title{Plot Labels for Counts & Proportion}
\usage{
n_percent(n, percent, reverse = FALSE)
}
\arguments{
\item{n}{Counts variable.}

\item{percent}{Proportion variable.}

\item{reverse}{Default set to false. Order switched to percent n if FALSE.}
}
\value{
Plot label as character.
}
\description{
Standardized plot labels with options for ordering.
}
\examples{
n_percent(5, 25)
n_percent(5, 25, reverse = TRUE)
}
