% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndromic.R
\name{nssp_date_range}
\alias{nssp_date_range}
\title{Adjust NSSP API Date Range}
\usage{
nssp_date_range(url, start = NULL, end = NULL)
}
\arguments{
\item{url}{ESSENCE API URL.}

\item{start}{Start date for data extraction, date format.}

\item{end}{End date for data extraction, date format.}
}
\value{
String API URL with new date range.
}
\description{
Customize date range in API URL without having to recreate within ESSENCE.
}
\examples{
\dontrun{
url <- "https::///www.syndromic.com/timeSeries?data_source&startDate=14Apr2024&medicalGroupingSystem&endDate=20Jul2024&percentParam"
nssp_date_range(url, start = as.Date("2024-01-01"), end = as.Date("2024-06-04"))
}
}
