% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cleveland.R, R/geom_lollipop.R,
%   R/highlight_geom.R
\docType{data}
\name{GeomCleveland}
\alias{GeomCleveland}
\alias{GeomLollipop}
\alias{ggplot_add.highlight}
\title{geom_cleveland proto}
\description{
geom_cleveland proto

geom_lollipop proto

ggplot_add.highlight
}
\keyword{datasets}
