% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calredie.R
\name{recode_gender}
\alias{recode_gender}
\title{Recode Gender/Sex in CalREDIE}
\usage{
recode_gender(col)
}
\arguments{
\item{col}{Input column from CalREDIE.}
}
\value{
Recoded input column.
}
\description{
Made specifically for CaLREDIE UDF data export. Recode abbreviations to full names.
}
\examples{
example = c("F","F","TF",NA)
recode_gender(example)
}
