% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calredie.R
\name{recode_race}
\alias{recode_race}
\title{Recode Race/Ethnicity}
\usage{
recode_race(ethnicity_col, race_col, abbr_names = FALSE)
}
\arguments{
\item{ethnicity_col}{Patient ethnicity variable.}

\item{race_col}{Patient race variable.}

\item{abbr_names}{TRUE/FALSE, option to abbreviate long category names.}
}
\value{
Merged race/ethnicity variable.
}
\description{
Simplify race/ethnicity patient data to one variable using CalREDIE, CAIR2, or VRBIS datasets. Hierarchy defaults to Hispanic/Latinx regardless of reported race. Function can handle one (e.g. CAIR2) or two inputs (e.g. CalREDIE). If using with VRBIS dataset, expected input is "Multi-status Race" column.
}
\examples{
recode_race("Hispanic","Black or African American")
recode_race("Native Hawaiian or Other Pacific Islander", abbr_names = FALSE)
recode_race("1")
}
