% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_metadata}
\alias{redcap_metadata}
\title{Read in Metadata from REDCap Project}
\usage{
redcap_metadata(url, token, forms = NULL, return = c("dictionary", "template"))
}
\arguments{
\item{url}{Character, API url.}

\item{token}{Character, token from project.}

\item{forms}{Vector, if not specified will include all forms.}

\item{return}{Character, dictionary or template. If bulk uploading new records, use template.}
}
\value{
Data.frame with project metadata.
}
\description{
This method allows you to read in the metadata for a project. To use this method, you must have API Export privileges in the project. If you're starting a new project and want to bulk upload records, you return = "template" to get a blank data.frame to populate. Otherwise, set return = "dictionary" to see all specs about project. You can send this dictionary to others to upload into their instance of REDCap.
}
