% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_cols.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove Empty Columns}
\usage{
remove_empty_cols(df)
}
\arguments{
\item{df}{Input dataframe.}
}
\value{
Dataframe with blank/empty columns removed. Message printed to console with total columns removed.
}
\description{
Ability to drop columns when either completely NA or "".
}
\examples{
test <- data.frame(a = c(NA,NA,NA), b = c("","",""), c = c(1,2,3))
remove_empty_cols(test)
}
