% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apollo.R
\name{theme_apollo}
\alias{theme_apollo}
\title{Apollo Theme}
\usage{
theme_apollo(
  direction = c("vertical", "horizontal", "map"),
  font = NULL,
  legend = "top"
)
}
\arguments{
\item{direction}{If using coord_flip, specify horizontal. Otherwise use vertical.}

\item{font}{Option to specify font otherwise set to default.}

\item{legend}{Set legend position to top, right, bottom, or left. To hide, use none.}
}
\value{
Theme for standard ggplots.
}
\description{
Apollo Theme
}
