% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_labels.R
\name{wrap_labels}
\alias{wrap_labels}
\title{Wrap Long Axis Labels}
\usage{
wrap_labels(width)
}
\arguments{
\item{width}{Higher value = less wrapping, lower value = more wrapping. Recommended width is 15.}
}
\value{
Wrapped labels.
}
\description{
Wrap long axis labels to avoid overcrowding.
}
\examples{
df <- data.frame(group = c("Native Hawaiian or Other Pacific Islander",
"Black or African American","American Indian or Alaska Native"), score = c(89.5, 84, 73))

ggplot(data = df, aes(x = group, y = score)) +
 geom_col() +
 scale_x_discrete(labels = wrap_labels(width = 15))
}
