% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{write_redcap}
\alias{write_redcap}
\title{Write Data to REDCap Project}
\usage{
write_redcap(df, url, token, content)
}
\arguments{
\item{df}{Data.frame to write back to REDCap.}

\item{url}{API url, character.}

\item{token}{API token, character.}

\item{content}{Currently limited to: record, userRole, user.}
}
\value{
Success or warning message, no data returned.
}
\description{
Write data from R back to REDCap project via API. You must be working with raw, unlabeled data.
}
